sync williamdev to main branch

+++ src/app/user-management/user-management.component.ts
@@ -90,14 +90,13 @@ export class UserManagementComponent {
       ctlRole: ['', Validators.required],
       ctlStatus: ['', Validators.required],
     });
-
-    this.editUserForm.get('ctlUsername')?.disable();
   }
 
   resetMessages(): void {
     this.error_msg.set(null);
     this.success_msg.set(null);
     this.newDialog_error_msg.set(null);
+    this.editDialog_error_msg.set(null);
   }
 
   filterUsers(event: Event): void {
@@ -175,7 +174,7 @@ export class UserManagementComponent {
     const newUser: User = {
       id: 0,
       username: this.newUserForm.value.ctlUsername,
-      password: 'password1', // Default password
+      password: '', // not in used
       email: this.newUserForm.value.ctlEmail,
       name: this.newUserForm.value.ctlName,
       roles: [this.newUserForm.value.ctlRole],
@@ -198,29 +197,13 @@ export class UserManagementComponent {
     });
   }
 
-  // confirmResetPassword(selectedUser: User): void {
-  //   this.resetMessages();
-  //   this.confirmationService.confirm({
-  //     message: 'Are you sure to reset password for user: <strong>' + selectedUser.username + '</strong>?',
-  //     header: 'Confirm Reset Password',
-  //     icon: 'pi pi-exclamation-triangle',
-  //     accept: () => {
-  //       // User confirmed, proceed with deletion
-  //       console.log('Resetting password for user:', selectedUser.username);
-  //     },
-  //     reject: () => {
-  //       // User rejected, do nothing
-  //       this.error_msg.set('Reset Password canceled.');
-  //       console.log('Reset Password canceled.');
-  //     }
-  //   });
-  // }
-
-  // resetPassword(selectedUser: User): void {
-  //   this.resetMessages();   
-  // }
-
   confirmDeleteUser(deletedUser: User): void {
+
+    if (deletedUser == null) {
+      this.error_msg.set('User not found');
+      return;
+    }
+
     this.resetMessages();
     this.confirmationService.confirm({
       message: 'Are you sure to delete user: <strong>' + deletedUser.username + '</strong>?',
@@ -239,6 +222,10 @@ export class UserManagementComponent {
   }
 
   deleteUser(deletedUser: User): void {
+    if (deletedUser == null) {
+      this.error_msg.set('User not found');
+      return;
+    }
     this.resetMessages();
     console.log(`Deleting user ${deletedUser.name}`);
     this.userService.deleteUser(deletedUser.id).subscribe({
@@ -259,6 +246,12 @@ export class UserManagementComponent {
   }
 
   showEditUserForm(user: User): void {
+
+    if (user === null) {
+      this.error_msg.set('User not found');
+      return;
+    }
+
     this.selectedUser.set(user);
     this.resetMessages();
     this.editUserForm.reset();
@@ -293,14 +286,19 @@ export class UserManagementComponent {
   }
 
   editUser(): void {
+
+    if (this.selectedUser() === null) {
+      this.error_msg.set('User not found');
+      return;
+    }
+
     this.resetMessages();
     console.log('Editing user');
-    console.log('User State:', this.editUserForm.value.ctlStatus === 'true');
 
     const editedUser: User = {
       id: this.selectedUser()?.id || 0,
       username: this.editUserForm.value.ctlUsername,
-      password: 'dummy', // not in used
+      password: '', // not in used
       email: this.editUserForm.value.ctlEmail,
       name: this.editUserForm.value.ctlName,
       roles: [this.editUserForm.value.ctlRole],
